refactor: extrait apiCall en fabrique partagee, migre le store clients
